rename to wfmock
